array_run: test for local process before cleaning temp folder, gave error in cluster

+++ cxsystem2/hpc/array_run.py
@@ -275,9 +275,10 @@ class ArrayRun:
         print(" -  Array run metadata saved at: %s" % os.path.join(
             os.path.dirname(paths[list(paths.keys())[0]]), self.metadata_filename))
 
-        tmp_folder_path = Path(parameter_finder(self.anatomy_df, 'workspace_path')).expanduser().joinpath('.tmp' + self.suffix).as_posix()
-        print("cleaning tmp folders " + tmp_folder_path)
-        shutil.rmtree(tmp_folder_path)
+        if self._is_running_locally() is True:
+            tmp_folder_path = Path(parameter_finder(self.anatomy_df, 'workspace_path')).expanduser().joinpath('.tmp' + self.suffix).as_posix()
+            print("cleaning tmp folders " + tmp_folder_path)
+            shutil.rmtree(tmp_folder_path)
 
     def generate_dataframes_for_param_search(self, original_df, index_of_array_variable, df_type, naming_prefix='', recursion_counter=1):
         """
